full_requirements.txt
# Complete ptools requirements - generated by scripts/generate_requirements.py.
# Do not edit by hand; regenerate with:
#   python scripts/generate_requirements.py > full-requirements.txt
#
# This file is intentionally NOT referenced from pyproject.toml - it exists
# only to make a full install reproducible in one shot.

# System binaries (install via your OS package manager):
#   - node OR deno OR bun

# --- Base dependencies (from pyproject.toml) ---
click>=8.1
humanize>=4.8.0
jinja2>=3.0.0
keyring>=25.6.0
lark>=1.1
numpy>=2.0.0
prompt-toolkit>=3.0.0
pycryptodome>=3.14.0
pydantic>=2.10
pyyaml>=6.0.0
requests>=2.32.0
textual>=0.1.0
watchdog>=3.0

# --- Optional libraries announced by ptools.utils.require ---
openai
pandas
pygments
pyperclip
